requirements.txt
streamlit
requests
google-generativeai
beautifulsoup4
wordcloud
textblob
matplotlib
nltk
pyyaml
networkx
pandas
numpy
praw